backend/requirements.txt
# ─── Web Framework ─────────────────────────────────────
fastapi==0.111.0
uvicorn[standard]==0.30.1
python-multipart==0.0.9
httpx==0.28.1

# ─── Auth ──────────────────────────────────────────────
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# ─── Database ──────────────────────────────────────────
sqlalchemy[asyncio]==2.0.30
asyncpg==0.29.0
alembic==1.13.1
psycopg2-binary==2.9.9
pgvector==0.4.2

# ─── Task Queue ────────────────────────────────────────
celery==5.4.0
redis==5.0.4
flower==2.0.1

# ─── GCP / Vertex AI ───────────────────────────────────
google-genai>=1.0.0
google-cloud-vision==3.7.2
google-cloud-storage>=2.16.0

# ─── Document Processing ───────────────────────────────
pymupdf==1.24.3
pdfplumber==0.11.0
python-docx==1.1.2
markdown==3.6
Pillow==10.3.0

# ─── Validation & Serialization ────────────────────────
pydantic==2.7.1
pydantic-settings==2.3.0

# ─── Utilities ─────────────────────────────────────────
python-dotenv==1.0.1
tenacity==8.3.0
structlog==24.2.0
sentry-sdk[fastapi]==2.3.1
prometheus-fastapi-instrumentator==7.0.0
slowapi==0.1.9
aiofiles==23.2.1
